Relocate JdbcQueryTranslator to a proper package
