+++ src/test/java/org/prebid/server/settings/JdbcApplicationSettingsTest.java
@@ -22,7 +22,7 @@ import org.prebid.server.exception.PreBidException;
 import org.prebid.server.execution.Timeout;
 import org.prebid.server.execution.TimeoutFactory;
 import org.prebid.server.metric.Metrics;
-import org.prebid.server.settings.helper.JdbcQueryTranslator;
+import org.prebid.server.settings.jdbc.JdbcQueryTranslator;
 import org.prebid.server.settings.model.Account;
 import org.prebid.server.settings.model.AccountAnalyticsConfig;
 import org.prebid.server.settings.model.AccountBidValidationConfig;
